+++ src/app/test/requests_test.py
@@ -23,6 +23,7 @@ from beer_garden.requests import (
     create_request,
     determine_latest_system_version,
     get_request,
+    rebroadcast,
 )
 from beer_garden.systems import create_system
 
@@ -1326,6 +1327,41 @@ class TestCancelRequest(object):
         cancel_request_children(request)
 
         cancel_mock.assert_called_once()
+
+
+class TestRebroadcastRequest(object):
+
+    @pytest.fixture(autouse=True)
+    def drop(self):
+        yield
+        Request.drop_collection()
+
+    @pytest.fixture
+    def test_request(self):
+        request = Request(
+            namespace="parent",
+            system="testsystem",
+            system_version="1.0.0",
+            instance_name="instance1",
+            command="somecommand",
+            parameters={},
+            status="SUCCESS",
+        )
+        request.save()
+
+        yield request
+
+    def test_rebroadcast_request(self, test_request, monkeypatch):
+        handle_event_rebroadcast_mock = Mock()
+        monkeypatch.setattr(
+            beer_garden.requests,
+            "handle_event_rebroadcast",
+            handle_event_rebroadcast_mock,
+        )
+
+        rebroadcast(test_request.id)
+
+        handle_event_rebroadcast_mock.assert_called_once()
 
 
 class TestCreateRequest(object):